+++ src/Spatter/CMakeLists.txt
@@ -1,7 +1,5 @@
 # Spatter/src/Spatter/CMakeLists.txt
 
-configure_file(Spatter_config.hpp.cmakein Spatter_config.hpp)
-
 if (USE_CUDA)
     add_library(cuda_backend SHARED CudaBackend.cu)
     set_target_properties(cuda_backend PROPERTIES CUDA_RUNTIME_LIBRARY Shared)
@@ -16,22 +14,27 @@ set(SPATTER_INCLUDE_FILES
     PatternParser.hh
     SpatterTypes.hh
     AlignedAllocator.hh
-    Timer.hh)
+    Timer.hh
+    )
 
-add_library(Spatter INTERFACE)
+add_library(Spatter
+    ${SPATTER_INCLUDE_FILES}
+    Configuration.cc
+    JSONParser.cc
+    PatternParser.cc
+    Timer.cc
+    )
 
 set_target_properties(Spatter PROPERTIES
-    INTERFACE_COMPILE_FEATURES cxx_std_20
     COMPILE_DEFINITIONS "${COMMON_COMPILE_DEFINITIONS}"
     COMPILE_OPTIONS "${COMMON_COMPILE_OPTIONS}"
     OUTPUT_NAME "Spatter"
     )
 
 target_include_directories(Spatter
-    INTERFACE
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}>
-    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+    PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>
+    $<INSTALL_INTERFACE:include/Spatter>
 )
 
 if (USE_CUDA)
@@ -39,27 +42,57 @@ if (USE_CUDA)
 endif()
 
 target_link_libraries(Spatter
-    INTERFACE
+    PUBLIC
     ${COMMON_LINK_LIBRARIES}
     )
 
 target_compile_options(Spatter
-    INTERFACE ${WARNING_FLAGS}
+    PRIVATE ${WARNING_FLAGS}
     )
 
+# Library/Header installation section
+
+#set(ConfigPackageLocation lib/cmake/Spatter)
+
 #install(TARGETS Spatter
-#    EXPORT Spatter_Targets
-#    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
-#    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
+#    EXPORT SpatterTargets
+#    LIBRARY DESTINATION lib
+#    ARCHIVE DESTINATION lib
+#    RUNTIME DESTINATION lib
+#    INCLUDES DESTINATION include
 #    )
 
-#install(EXPORT Spatter_Targets
-#    FILE Spatter_Targets.cmake
+#install(FILES ${SPATTER_INCLUDE_FILES}
+#    DESTINATION include
+#    )
+
+#install(EXPORT SpatterTargets
+#    FILE SpatterTargets.cmake
 #    NAMESPACE Spatter::
-#    DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/Spatter
+#    DESTINATION ${ConfigPackageLocation}
+#    )
+
+#include(CMakePackageConfigHelpers)
+
+#write_basic_package_version_file(
+#    SpatterConfigVersion.cmake
+#    VERSION ${PACKAGE_VERSION}
+#    COMPATIBILITY AnyNewerVersion
 #)
 
-#install(FILES ${SPATTER_INCLUDE_FILES} DESTINATION ${CMAKE_INSTALL_INCLUDEDIR})
+#configure_package_config_file(${CMAKE_CURRENT_SOURCE_DIR}/Config.cmake.in
+#    "${CMAKE_CURRENT_BINARY_DIR}/SpatterConfig.cmake"
+#    INSTALL_DESTINATION ${ConfigPackageLocation}
+#    )
 
-#install(FILES ${CMAKE_CURRENT_BINARY_DIR}/Spatter_config.hpp DESTINATION ${CMAKE_INSTALL_INCLUDEDIR})
+#configure_file(SpatterConfig.cmake.in Spatter.cmake @ONLY)
 
+#export(EXPORT SpatterTargets
+#       FILE "${CMAKE_CURRENT_BINARY_DIR}/cmake/SpatterTargets.cmake"
+#       NAMESPACE Spatter::
+#)
+
+#install(FILES "${CMAKE_CURRENT_BINARY_DIR}/SpatterConfig.cmake"
+#              "${CMAKE_CURRENT_BINARY_DIR}/SpatterConfigVersion.cmake"
+#        DESTINATION ${ConfigPackageLocation}
+#        )
