+++ src/Spatter/Config.cmake.in
@@ -1,0 +1,5 @@
+@PACKAGE_INIT@
+
+include("${CMAKE_CURRENT_LIST_DIR}/SpatterTargets.cmake")
+
+check_required_components(Spatter)
