+++ src/Spatter/Timer.hh
@@ -16,22 +16,10 @@ class Timer {
   typedef std::chrono::system_clock CLOCK_T;
 
 public:
-  void start() {
-    assert(!running);
-    running = true;
-    start_tp = CLOCK_T::now();
-  }
-  void stop() {
-    std::chrono::time_point<CLOCK_T> stop_tp = CLOCK_T::now();
-    assert(running);
-    running = false;
-    accum += stop_tp - start_tp;
-  }
-  double seconds() const { return accum.count(); }
-  void clear() {
-    running = false;
-    accum = std::chrono::duration<double>(0.0);
-  }
+  void start();
+  void stop();
+  double seconds() const;
+  void clear();
 
 private:
   bool running = false;
@@ -39,9 +27,7 @@ private:
   std::chrono::time_point<CLOCK_T> start_tp;
 };
 
-inline std::ostream &operator<<(std::ostream &os, const Timer &t) {
-  return os << t.seconds() << 's';
-}
+inline std::ostream &operator<<(std::ostream &os, const Timer &t);
 
 } // namespace Spatter
 #endif
